Massive refactor of actor/policy class

+++ src/data/process_demos.py
@@ -81,9 +81,9 @@ def process_demos_to_feature(input_path, output_path, encoder, batch_size=256):
         features_list2.append(features2)
 
     # Convert to numpy arrays
-    observations = np.array(observations)
-    actions = np.array(actions)
-    episode_ends = np.array(episode_ends)
+    observations = np.array(observations, dtype=np.float32)
+    actions = np.array(actions, dtype=np.float32)
+    episode_ends = np.array(episode_ends, dtype=np.float32)
 
     # Save to file
     output_path.mkdir(parents=True, exist_ok=True)
